controllers

+++ AcademyApp/Controllers/GroupController.cs
@@ -22,6 +22,7 @@ namespace Manage.Controllers
         {
             ConsoleHelper.WriteTextWithColor(ConsoleColor.Magenta, "Enter group name:");
             string name = Console.ReadLine();
+
         MaxSize: ConsoleHelper.WriteTextWithColor(ConsoleColor.Magenta, "Enter group max size:");
             string size = Console.ReadLine();
             int maxSize;
@@ -47,33 +48,66 @@ namespace Manage.Controllers
         #region update
         public void UpdateGroup()
         {
-            ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkBlue, "Enter group name:");
-            string name = Console.ReadLine();
-
-            var group = _groupRepository.Get(g => g.Name.ToLower() == name.ToLower());
-            if (group != null)
+            var groups = _groupRepository.GetAll();
+            if (groups.Count > 0)
             {
-                int oldSize = group.MaxSize;
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkBlue, "Enter new group name:");
-                string newName = Console.ReadLine();
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.Yellow, "All groups");
 
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkBlue, "Enter new group max size:");
-                string size = Console.ReadLine();
+                foreach (var dbGroup in groups)
+                {
+                    Console.WriteLine($"Id-{dbGroup.Id}, name-{dbGroup.Name}");
+                }
+            ID: ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkBlue, "Enter group id:");
+                int chosenId;
+                string id = Console.ReadLine();
+                var result = int.TryParse(id, out chosenId);
 
-                int maxSize;
-                bool result = int.TryParse(size, out maxSize);
                 if (result)
                 {
-                    var newGroup = new Group
+                    var group = _groupRepository.Get(g => g.Id == chosenId);
+                    if (group != null)
                     {
-                        Id = group.Id,
-                        Name = newName,
-                        MaxSize = maxSize
-                    };
-                    _groupRepository.Update(newGroup);
-                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Blue, $"{newGroup.Name} is new name, {newGroup.MaxSize} is new max size");
-                }
+                        int oldSize = group.MaxSize;
+                        string oldName = group.Name;
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkBlue, "Enter new group name:");
+                        string newName = Console.ReadLine();
 
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.DarkBlue, "Enter new group max size:");
+                        string size = Console.ReadLine();
+
+                        int maxSize;
+                        result = int.TryParse(size, out maxSize);
+
+                        if (result)
+                        {
+                            var newGroup = new Group
+                            {
+                                Id = group.Id,
+                                Name = newName,
+                                MaxSize = maxSize
+                            };
+                            _groupRepository.Update(newGroup);
+                            ConsoleHelper.WriteTextWithColor(ConsoleColor.Blue, $"Name:{oldName} Max size:, {oldSize} is updated to name:");
+                        }
+                        else
+                        {
+                            ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, $"Please enter correct group max size:");
+                        }
+                    }
+                    else
+                    {
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, $"Please enter correct group name:");
+                    }
+                }
+                else
+                {
+                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, $"Please enter correct id:");
+                    goto ID;
+                }
+            }
+            else
+            {
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, $"There are no any groups:");
             }
         }
         #endregion
@@ -117,35 +151,55 @@ namespace Manage.Controllers
 
         public void DeleteGroup()
         {
-            ConsoleHelper.WriteTextWithColor(ConsoleColor.Yellow, "Enter group name");
-            string name = Console.ReadLine();
-            var group = _groupRepository.Get(g => g.Name.ToLower() == name.ToLower());
-            if (group != null)
+            var groups = _groupRepository.GetAll();
+            if (groups.Count > 0)
             {
-                _groupRepository.Delete(group);
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.Gray, $"{name} is deleted");
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.Yellow, "All groups");
 
+                foreach (var dbGroup in groups)
+                {
+                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Magenta, $"Id - {dbGroup.Id}, Name -{dbGroup.Name}");
+                }
+            Id: ConsoleHelper.WriteTextWithColor(ConsoleColor.Yellow, "Enter group id:");
+                int chosenId;
+                string id = Console.ReadLine();
+                var result = int.TryParse(id, out chosenId);
+                if (result)
+                {
+                    var group = _groupRepository.Get(g => g.Id == chosenId);
+                    if (group != null)
+                    {
+                        string name = group.Name;
+                        _groupRepository.Delete(group);
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.Green, $"{name} is deleted");
+                    }
+                    else
+                    {
+                        ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "This group doesn't exist");
+
+                    }
+                }
+                else
+                {
+                    ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "Please, enter correct Id");
+                    goto Id;
+                }
             }
             else
             {
-                ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "This group doesn't exist");
+                ConsoleHelper.WriteTextWithColor(ConsoleColor.Red, "There are no any groups");
             }
-
-
-
         }
-
-        #endregion 
-
+        #endregion
 
         #region Exit
 
         public void Exit()
-    {
-        ConsoleHelper.WriteTextWithColor(ConsoleColor.Green, "Thanks for using our application");
+        {
+            ConsoleHelper.WriteTextWithColor(ConsoleColor.Green, "Thanks for using our application");
 
+        }
+
+        #endregion
     }
-
-    #endregion
-     }
 }
